+++ src/ServiceProvider.php
@@ -3,7 +3,7 @@
 namespace Johncarter\StatamicLivewireForms;
 
 use Johncarter\StatamicLivewireForms\Commands\MakeViewCommand;
-use Johncarter\StatamicLivewireForms\Http\Livewire\LivewireForm;
+use Johncarter\StatamicLivewireForms\Http\Livewire\StatamicForm;
 use Livewire\Livewire;
 use Statamic\Providers\AddonServiceProvider;
 
@@ -12,7 +12,7 @@ class ServiceProvider extends AddonServiceProvider
     public function boot()
     {
         parent::boot();
-        Livewire::component('livewire-form', LivewireForm::class);
+        Livewire::component('statamic-form', StatamicForm::class);
 
         if ($this->app->runningInConsole()) {
             $this->commands([